Added popular places in the community (#3)
* Implemented simple interative renderer using matplotlib

 - Used the matplotlib animation API to render a community
 - Added sliders in the render to control two attributes of the
    population
 - Added a button to pause the simulation at any point
 - The simulation can be run indefinitely or for certain number of steps

* Added a sample requirements file, also updated gitignore

* Added popular places in the community

The Community class now accepts a list of popular places which the
people frequently visit with a probability. The people might either go
to one of the popular places or randomly visit a point depending on the
probability parameter specified. Also added a slider for this parameter
in the render.

* Added a sample requirements file, also updated gitignore

* Added popular places in the community

The Community class now accepts a list of popular places which the
people frequently visit with a probability. The people might either go
to one of the popular places or randomly visit a point depending on the
probability parameter specified. Also added a slider for this parameter
in the render.

* Better walk range calculation